pyproject.toml
[tool.poetry]
name = "large-gcs"
version = "0.1.0"
description = ""
authors = ["Shao Yuan <flossy_lineage.0b@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
numpy = "^1.23.5"
matplotlib = "^3.7.1"
ipykernel = "^6.23.1"
scipy = "^1.10.1"
graphviz = "^0.20.1"
black = "^24.3.0"
tqdm = "^4.65.0"
pytest = "^7.4.0"
wandb = "^0.15.5"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
autoflake = "^2.2.0"
isort = "^5.12.0"
pdbpp = "^0.10.3"
plotly = "^5.20.0"
docformatter = "^1.7.5"
drake = "^1.28.0"
manipulation = "^2024.5.25"
nbformat = "^5.10.4"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
addopts = "--disable-warnings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

